refactor and add enhanced error handling for startwallpapers

+++ components/tasks/GetNextBackground.brs
@@ -6,13 +6,16 @@ end sub
 
 sub GetNextBackground()
 
-    print "Running Get Next Background"
     fs = CreateObject("roFileSystem")
     fs.Delete(m.global.backgroundUri)
 
-    ' if m.global.imgIndex >= m.global.keyList.Count()
-    '     m.global.imgIndex = 0
-    ' end if
+    if m.global.resourceLinks = invalid or m.global.keyList = invalid
+        m.top.result = "fail"
+        return
+    else if m.global.resourceLinks.Count() = 0 or m.global.keyList.Count() = 0
+        m.top.result = "fail"
+        return
+    end if
 
     m.currHeader = {
         "Authorization": m.global.resourceLinks[m.global.keyList[m.global.imgIndex]],
@@ -24,23 +27,22 @@ sub GetNextBackground()
 
     m.imageHttp = CreateObject("roUrlTransfer")
     m.imageHttp.SetHeaders(m.currHeader)
-    m.imageHttp.SetUrl("http://10.0.0.15:8080/link/background")
+    m.imageHttp.SetUrl(m.global.webappUrl + "/link/background")
     m.imgHttpPort = CreateObject("roMessagePort")
     m.imageHttp.SetPort(m.imgHttpPort)
     m.imageHttp.AsyncGetToFile("tmp:/" + m.global.filenameCounter.ToStr())
     m.imgResponse = Wait(5000, m.imgHttpPort)
 
-    ' handle if the keys did not work at the web app
+    if m.imgResponse = invalid
+        m.top.result = "fail"
+        return
+    end if
 
     m.responseCode = m.imgResponse.GetResponseCode()
 
-    if m.responseCode <> invalid
-        print "response code valid " + m.responseCode.ToStr()
-        if m.responseCode = 401
-            print "response code equals 401"
-            m.top.result = "401"
-            return
-        end if
+    if m.responseCode = 401
+        m.top.result = "keyFail"
+        return
     end if
 
     if m.responseCode <> 200
@@ -48,30 +50,26 @@ sub GetNextBackground()
         return
     end if
 
-    if m.imgResponse <> invalid
-        m.responseHeaders = m.imgResponse.GetResponseHeaders()
-        for each item in m.responseHeaders
-            if item = "content-type"
-                m.fileType = m.responseHeaders[item]
-                m.cleanType = Mid(m.fileType, Len("image/j"))
-            end if
-        end for
+    m.responseHeaders = m.imgResponse.GetResponseHeaders()
+
+    if m.responseHeaders.Lookup("content-type") <> invalid
+        m.fileType = m.responseHeaders["content-type"]
+        m.cleanType = "." + Mid(m.fileType, Len("image/j"))
+        if m.cleanType = "."
+            m.cleanType = ""
+        end if
+    else
+        m.cleanType = ""
+    end if
+
+    m.finalImgName = "tmp:/background" + m.global.filenameCounter.ToStr() + m.cleanType
+
+    if fs.CopyFile("tmp:/" + m.global.filenameCounter.ToStr(), m.finalImgName)
+        fs.Delete("tmp:/" + m.global.filenameCounter.ToStr())
+        m.global.backgroundUri = m.finalImgName
+        m.top.result = m.finalImgName
     else
         m.top.result = "fail"
         return
     end if
-
-    m.finalImgName = "tmp:/background" + m.global.filenameCounter.ToStr() + "." + m.cleanType
-
-    success = fs.CopyFile("tmp:/" + m.global.filenameCounter.ToStr(), m.finalImgName)
-    fs.Delete("tmp:/" + m.global.filenameCounter.ToStr())
-
-    if success
-        m.global.backgroundUri = m.finalImgName
-        m.top.result = m.finalImgName
-    else
-        print "failed to rename the background file"
-    end if
-
-    print "Finish get next background"
 end sub
